+++ apps/chat/tests/test_openapi_tool.py
